mejoras

+++ database/migrations/2019_10_27_174106_create_estilo_aprendizaje_table.php
@@ -15,7 +15,7 @@ class CreateEstiloAprendizajeTable extends Migration
     {
         Schema::create('estilo_aprendizaje', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('user_id')->unsigned();            
+            $table->bigInteger('user_id')->unsigned();            
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->string('estilo');
             $table->timestamps();
