+++ database/migrations/2019_10_27_193524_create_pagina_table.php
@@ -15,7 +15,7 @@ class CreatePaginaTable extends Migration
     {
         Schema::create('pagina', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('recurso_id')->unsigned();            
+            $table->bigInteger('recurso_id')->unsigned();            
             $table->foreign('recurso_id')->references('id')->on('recurso_instruccional')->onDelete('cascade');
             $table->string('descripcion');
             $table->string('formato');
