+++ database/migrations/2019_10_27_200401_create_dispositivo_disponible_table.php
@@ -15,14 +15,14 @@ class CreateDispositivoDisponibleTable extends Migration
     {
         Schema::create('dispositivo_disponible', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('componente_virtual_id')->unsigned();            
+            $table->bigInteger('componente_virtual_id')->unsigned();            
             $table->foreign('componente_virtual_id')->references('id')->on('componente_virtual')->onDelete('cascade');
-            $table->integer('ancho_banda')->unsigned();
+            $table->bigInteger('ancho_banda')->unsigned();
             $table->boolean('audio');
-            $table->integer('memoria_ram')->unsigned();
-            $table->integer('memoria_secundaria')->unsigned();
+            $table->bigInteger('memoria_ram')->unsigned();
+            $table->bigInteger('memoria_secundaria')->unsigned();
             $table->string('resolucion_pantalla');
-            $table->integer('uso_procesador')->unsigned();
+            $table->bigInteger('uso_procesador')->unsigned();
             $table->timestamps();
         });
     }
