+++ database/migrations/2019_10_27_200907_create_equipamento_table.php
@@ -15,12 +15,12 @@ class CreateEquipamentoTable extends Migration
     {
         Schema::create('equipamento', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('componente_fisico_id')->unsigned();            
+            $table->bigInteger('componente_fisico_id')->unsigned();            
             $table->foreign('componente_fisico_id')->references('id')->on('componente_fisico')->onDelete('cascade');
             $table->boolean('audio');
             $table->boolean('gps');
-            $table->integer('memoria_ram')->unsigned();
-            $table->integer('memoria_secundaria')->unsigned();
+            $table->bigInteger('memoria_ram')->unsigned();
+            $table->bigInteger('memoria_secundaria')->unsigned();
             $table->string('pantalla');
             $table->string('procesador');
             $table->string('tipo_disponible_acceso');
