+++ database/migrations/2019_10_27_201230_create_localizacion_table.php
@@ -15,7 +15,7 @@ class CreateLocalizacionTable extends Migration
     {
         Schema::create('localizacion', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('componente_fisico_id')->unsigned();            
+            $table->bigInteger('componente_fisico_id')->unsigned();            
             $table->foreign('componente_fisico_id')->references('id')->on('componente_fisico')->onDelete('cascade');
             $table->string('ciudad');
             $table->string('idioma');
