+++ database/migrations/2019_10_29_150622_create_lom_educativo_table.php
@@ -15,7 +15,7 @@ class CreateLomEducativoTable extends Migration
     {
         Schema::create('lom_educativo', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('oa_id')->unsigned();            
+            $table->bigInteger('oa_id')->unsigned();            
             $table->foreign('oa_id')->references('id')->on('oa')->onDelete('cascade');
             $table->string('contexto_uso');
             $table->string('densidad_semantica');
