+++ database/migrations/2019_10_29_150803_create_lom_tecnica_table.php
@@ -15,7 +15,7 @@ class CreateLomTecnicaTable extends Migration
     {
         Schema::create('lom_tecnica', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('oa_id')->unsigned();            
+            $table->bigInteger('oa_id')->unsigned();            
             $table->foreign('oa_id')->references('id')->on('oa')->onDelete('cascade');
             $table->string('duracion');
             $table->string('formato');
